WIP: improve flexibility of Jellyfin publishing
In accordance with our discussion and with help from JanV I have
finally managed to put together the composite build for Jellyfin.
I haven't yet had a chance to test whether it actually runs, I
hope to get to it in the next 1-2 days, but it should be building
the right components.

Thanks

Tomas